pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astrobio-gen"
version = "1.0.0"
description = "World-Class Astrobiology Research Platform with AGI Capabilities"
readme = "README.md"
license = {file = "LICENSE.md"}
authors = [
    {name = "Astrobio Research Team"},
]
maintainers = [
    {name = "Astrobio Research Team"},
]
keywords = [
    "astrobiology",
    "exoplanets",
    "machine-learning",
    "agi",
    "scientific-discovery",
    "autonomous-research",
    "observatories",
    "climate-modeling",
    "multimodal-ai",
    "causal-inference"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
requires-python = ">=3.9"
dependencies = [
    # Core ML frameworks
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "lightning>=2.0.0",
    
    # Scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "xarray>=2023.1.0",
    "zarr>=2.14.0",
    
    # Astronomy
    "astropy>=5.2.0",
    "astroquery>=0.4.6",
    
    # Data processing
    "h5py>=3.8.0",
    "netcdf4>=1.6.2",
    "dask[complete]>=2023.1.0",
    
    # ML extras
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "peft>=0.15.0",
    "datasets>=2.12.0",
    
    # Graph networks
    "torch-geometric>=2.3.0",
    "torch-scatter>=2.1.0",
    "torch-sparse>=0.6.17",
    
    # Configuration and experiment tracking
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    "mlflow>=2.4.0",
    
    # Web and API
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "streamlit>=1.25.0",
    "gradio>=3.35.0",
    
    # Utilities
    "rich>=13.4.0",
    "tqdm>=4.65.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    
    # Development
    "jupyter>=1.0.0",
    "notebook>=6.5.0",
    "ipywidgets>=8.0.0",
]

[project.optional-dependencies]
gpu = [
    "nvidia-ml-py3>=7.352.0",
    "tritonclient[all]>=2.40.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.1.0",
    "mkdocstrings[python]>=0.22.0",
]
all = [
    "astrobio-gen[gpu,dev,docs]",
]

[project.scripts]
astro-train = "astrobio_gen.cli:train_cli"
astro-serve = "astrobio_gen.cli:serve_cli" 
astro-data = "astrobio_gen.cli:data_cli"
astro-eval = "astrobio_gen.cli:eval_cli"

[project.urls]
Homepage = "https://github.com/astrobio-research/astrobio-gen"
Documentation = "https://astrobio-gen.readthedocs.io"
Repository = "https://github.com/astrobio-research/astrobio-gen"
"Bug Tracker" = "https://github.com/astrobio-research/astrobio-gen/issues"

[tool.setuptools]
packages = ["astrobio_gen"]

[tool.setuptools.package-dir]
astrobio_gen = "src/astrobio_gen"

[tool.setuptools.package-data]
astrobio_gen = [
    "config/*.yaml",
    "config/**/*.yaml", 
    "data/*.json",
    "models/*.pt",
    "models/*.pth",
]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  __pycache__|
  \.git|
  \.mypy_cache|
  \.pytest_cache|
  \.venv|
  astrobio_venv|
  _build|
  buck-out|
  build|
  dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["B011"]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "lightning.*",
    "transformers.*",
    "astropy.*",
    "wandb.*",
    "mlflow.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=astrobio_gen",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["astrobio_gen"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
